pyproject.toml
[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[project]
name = "docugen"
version = "0.1.0"
description = "MCP server: PDF spec → chaptered documentary video"
requires-python = ">=3.10"
dependencies = [
    "mcp[cli]>=1.0.0",
    "openai>=1.0.0",
    "manim>=0.18.0",
    "manimpango>=0.5.0",
    "pymupdf>=1.24.0",
    "numpy>=1.24.0",
    "scipy>=1.10.0",
    "pyyaml>=6.0",
    "chatterbox-mlx>=1.0.4",
    "torchaudio>=2.8.0",
    "pdf2image>=1.17.0",
    "openai-whisper>=20250625",
    "setuptools<81",
]

[project.optional-dependencies]
dev = ["pytest>=7.0"]

[project.scripts]
docugen = "docugen.server:main"

[tool.setuptools.packages.find]
where = ["src"]
